Refactor equipment to weapon sets with off-hands
Reworked player equipment from 6 slots to 8 slots by replacing legacy weapon slots with Set A/Set B main-hand + off-hand, introducing `WeaponSetSlot` and weapon handedness (`OneHanded`/`TwoHanded`). Updated equip/mutation rules, input actions, active-weapon selection, raid admission payloads (codec v9), and local profile persistence (schema v4) to carry the new model and validate two-handed/off-hand constraints. UI/prefabs, stash/town flows, tests, and architecture docs were aligned to explicit slot targeting and blocked off-hand behavior.

+++ Project Grimhold/Assets/Scripts/Player/PlayerRaidLootOriginState.cs
@@ -10,9 +10,9 @@ public sealed class PlayerRaidLootOriginState : NetworkBehaviour
     [Networked]
     private RaidLootOriginPackedState InventoryOrigins { get; set; }
 
-    // Six 5-bit stable origin slots, one for each EquipmentSlotRules.AllSlots entry.
+    // Eight 5-bit stable origin slots, one for each EquipmentSlotRules.AllSlots entry.
     [Networked]
-    private int EquipmentOriginSlots { get; set; }
+    private long EquipmentOriginSlots { get; set; }
 
     public bool TryInitializePlayerLoadout(
         IReadOnlyList<LootEntry> entries,
@@ -208,13 +208,13 @@ public sealed class PlayerRaidLootOriginState : NetworkBehaviour
     }
 
     private int GetEquipmentOriginSlot(int equipmentIndex) =>
-        (EquipmentOriginSlots >> (equipmentIndex * 5)) & 31;
+        (int)((EquipmentOriginSlots >> (equipmentIndex * 5)) & 31L);
 
     private void SetEquipmentOriginSlot(int equipmentIndex, int originSlot)
     {
         int shift = equipmentIndex * 5;
-        int mask = 31 << shift;
-        EquipmentOriginSlots = (EquipmentOriginSlots & ~mask) | (originSlot << shift);
+        long mask = 31L << shift;
+        EquipmentOriginSlots = (EquipmentOriginSlots & ~mask) | ((long)originSlot << shift);
     }
 
     private static bool TryGetEquipmentIndex(EquipmentSlot slot, out int equipmentIndex)
